remove previous google map wrapper implementation
